refactor (bitbucket): resolve commits
