+++ src/app/app.component.html
@@ -1,5 +1,2 @@
-<app-header></app-header>
-<app-breadcrumbs></app-breadcrumbs>
-<app-section></app-section>
-<app-courses-list></app-courses-list>
-<app-footer></app-footer>
+<app-dashboard-layout></app-dashboard-layout>
+<router-outlet></router-outlet>
