--- src/app/breadcrumbs/breadcrumbs.component.html
@@ -1,1 +1,0 @@
-<p>breadcrumbs works!</p>
