--- src/app/breadcrumbs/breadcrumbs.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-
-import { BreadcrumbsComponent } from "./breadcrumbs.component";
-
-describe("BreadcrumbsComponent", () => {
-  let component: BreadcrumbsComponent;
-  let fixture: ComponentFixture<BreadcrumbsComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ BreadcrumbsComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(BreadcrumbsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it("should create", () => {
-    expect(component).toBeTruthy();
-  });
-});
