--- src/app/breadcrumbs/breadcrumbs.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-  selector: "app-breadcrumbs",
-  templateUrl: "./breadcrumbs.component.html",
-  styleUrls: ["./breadcrumbs.component.scss"]
-})
-export class BreadcrumbsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
