--- src/app/breadcrumbs/index.ts
@@ -1,1 +1,0 @@
-export * from "./breadcrumbs.component";
