+++ src/app/core/models/course-item.model.ts
@@ -2,62 +2,28 @@
  * CourseItem interface
  */
 export interface CourseItemModel {
-  // id of course item
-  id: number;
-
-  // title of course
-  title: string;
-
-  // creation date
-  creationDate: Date | number;
-
-  // course's duration
-  duration: number;
-
-  // description for course item
-  description: string;
-}
-
-/**
- * CourseItem class that implements CourseItem iterface
- */
-export class CourseItem implements CourseItemModel {
   /**
    * id of course item
    */
-  public id: number;
+  id: number;
 
   /**
    * title of course
    */
-  public title: string;
+  title: string;
 
   /**
    * creation date
    */
-  public creationDate: Date | number;
+  creationDate: Date | number;
 
   /**
    * course's duration
    */
-  public duration: number;
+  duration: number;
 
   /**
    * description for course item
    */
-  public description: string;
-
-  constructor(
-    id: number,
-    title: string,
-    creationDate: Date | number,
-    duration: number,
-    description: string,
-  ) {
-    this.id = id;
-    this.title = title;
-    this.creationDate = creationDate;
-    this.duration = duration;
-    this.description = description;
-  }
+  description: string;
 }
