+++ src/app/core/models/index.ts
@@ -1,2 +1,4 @@
-export { CourseItemModel, CourseItem } from "./course-item.model";
-export { UserEntityModel, UserEntity } from "./user-entity.model";
+export * from "./course-item.model";
+export * from "./user-entity.model";
+export * from "./course-item";
+export * from "./user-entity";
