+++ src/app/core/models/user-entity.model.ts
@@ -2,38 +2,18 @@
  *  UserEntity interface
  */
 export interface UserEntityModel {
-  // id of course item
-  id: number;
-
-  // user's first name
-  firstName: string;
-
-  // user's last name
-  lastName: string;
-}
-
-/**
- * UserEntity class that implements UserEntity iterface
- */
-export class UserEntity implements UserEntityModel {
   /**
    * id of course item
    */
-  public id: number;
+  id: number;
 
   /**
    * user's first name
    */
-  public firstName: string;
+  firstName: string;
 
   /**
    * user's last name
    */
-  public lastName: string;
-
-  constructor(id: number, firstName: string, lastName: string) {
-    this.id = id;
-    this.firstName = firstName;
-    this.lastName = lastName;
-  }
+  lastName: string;
 }
