+++ src/app/dashboard-layout/courses-list/course-item/course-item.component.html
