+++ src/app/dashboard-layout/courses-list/course-item/course-item.component.ts
@@ -6,7 +6,5 @@ import { Component, OnInit } from "@angular/core";
   styleUrls: ["./course-item.component.scss"],
 })
 export class CourseItemComponent implements OnInit {
-  constructor() {}
-
   ngOnInit() {}
 }
