+++ src/app/dashboard-layout/courses-list/course-item/index.ts
