+++ src/app/dashboard-layout/courses-list/courses-list.component.ts
@@ -3,13 +3,8 @@ import { Component, OnInit } from "@angular/core";
 @Component({
   selector: "app-courses-list",
   templateUrl: "./courses-list.component.html",
-  styleUrls: ["./courses-list.component.scss"]
+  styleUrls: ["./courses-list.component.scss"],
 })
 export class CoursesListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+  ngOnInit() {}
 }
