+++ src/app/dashboard-layout/courses-list/index.ts
