+++ src/app/dashboard-layout/dashboard-layout.component.html
@@ -1,0 +1,4 @@
+<p>dashboard-layout works!</p>
+<app-header></app-header>
+<app-courses-list></app-courses-list>
+<app-footer></app-footer>
