+++ src/app/dashboard-layout/dashboard-layout.component.spec.ts
@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { SectionComponent } from "./section.component";
+import { DashboardLayoutComponent } from "./dashboard-layout.component";
 
-describe("SectionComponent", () => {
-  let component: SectionComponent;
-  let fixture: ComponentFixture<SectionComponent>;
+describe("DashboardLayoutComponent", () => {
+  let component: DashboardLayoutComponent;
+  let fixture: ComponentFixture<DashboardLayoutComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ SectionComponent ]
+      declarations: [ DashboardLayoutComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(SectionComponent);
+    fixture = TestBed.createComponent(DashboardLayoutComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
