+++ src/app/dashboard-layout/footer/footer.component.ts
@@ -3,13 +3,8 @@ import { Component, OnInit } from "@angular/core";
 @Component({
   selector: "app-footer",
   templateUrl: "./footer.component.html",
-  styleUrls: ["./footer.component.scss"]
+  styleUrls: ["./footer.component.scss"],
 })
 export class FooterComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+  ngOnInit() {}
 }
