+++ src/app/dashboard-layout/footer/index.ts
