+++ src/app/dashboard-layout/header/logo/logo.component.ts
@@ -3,13 +3,8 @@ import { Component, OnInit } from "@angular/core";
 @Component({
   selector: "app-logo",
   templateUrl: "./logo.component.html",
-  styleUrls: ["./logo.component.scss"]
+  styleUrls: ["./logo.component.scss"],
 })
 export class LogoComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
+  ngOnInit() {}
 }
