+++ src/app/dashboard-layout/index.ts
@@ -1,0 +1,5 @@
+export * from "./dashboard-layout.component";
+export * from "./courses-list/index";
+export * from "./footer/index";
+export * from "./header/index";
+export * from "./dashboard-layout.module";
