--- src/app/section/index.ts
@@ -1,1 +1,0 @@
-export * from "./section.component";
