--- src/app/section/section.component.html
@@ -1,1 +1,0 @@
-<p>section works!</p>
