--- src/app/section/section.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from "@angular/core";
-
-@Component({
-  selector: "app-section",
-  templateUrl: "./section.component.html",
-  styleUrls: ["./section.component.scss"]
-})
-export class SectionComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
